quantum_flow_bybit.py
"""
QuantumFlow Elite Bybit - Micro Account Edition
Last Updated: 2025-03-29 08:50:23 UTC
User: Ikaraoha
Version: 2.1.0
"""

import os
import logging
import time
from datetime import datetime, timedelta
from typing import Dict, List, Optional
from pybit.unified_trading import HTTP
from telegram import Update
from telegram.ext import Application, CommandHandler, ContextTypes
import pandas as pd
import numpy as np
from config import (SETTINGS, PAIRS, RECOVERY_SETTINGS, 
                   COMPOUND_SETTINGS, TRADE_LIMIT_TIERS)

class QuantumFlowBybit:
    def __init__(self):
        self.timestamp = "2025-03-29 08:50:23"
        self.user = "Ikaraoha"
        self.version = "2.1.0"
        self.logger = self.setup_logging()
        
        # Load configurations
        self.settings = SETTINGS
        self.pairs = PAIRS
        self.recovery = RECOVERY_SETTINGS
        self.compound = COMPOUND_SETTINGS
        
        # Initialize tracking variables
        self.trades_today = {pair: 0 for pair in PAIRS}
        self.daily_profit = 0.0
        self.consecutive_losses = 0
        self.compound_level = 1
        self.last_trade_time = {}
        self.recovery_mode = False
        self.current_trade_limits = {}
        self.last_limit_update = datetime.utcnow()
        
        # Initialize Bybit client
        self.client = HTTP(
            testnet=False,
            api_key=os.getenv('BYBIT_MAINNET_API_KEY'),
            api_secret=os.getenv('BYBIT_MAINNET_API_SECRET')
        )
        
        # Perform startup checks
        self.initial_balance = self.get_account_balance()
        self.perform_startup_checks()
        self.current_trade_limits = self.calculate_trade_limits()

    def calculate_trade_limits(self) -> Dict[str, int]:
        """Calculate dynamic trade limits based on current balance"""
        try:
            current_balance = self.get_account_balance()
            
            # Find appropriate tier
            max_trades = TRADE_LIMIT_TIERS[10]  # Default to lowest tier
            for balance_threshold in sorted(TRADE_LIMIT_TIERS.keys()):
                if current_balance >= balance_threshold:
                    max_trades = TRADE_LIMIT_TIERS[balance_threshold]
                else:
                    break
            
            # Calculate limits for each pair
            limits = {}
            for pair, settings in self.pairs.items():
                base_trades = settings['risk_settings']['base_daily_trades']
                scaled_trades = int(base_trades * (max_trades / 4))
                limits[pair] = scaled_trades
            
            # Log new limits
            self.logger.info(f"New trade limits calculated: {limits}")
            return limits
            
        except Exception as e:
            self.logger.error(f"Trade limit calculation error: {str(e)}")
            return {pair: settings['risk_settings']['base_daily_trades'] 
                   for pair, settings in self.pairs.items()}

    def calculate_adaptive_lot(self, symbol: str, base_lot: float) -> float:
        """Enhanced adaptive lot size calculation"""
        try:
            current_balance = self.get_account_balance()
            multiplier = 1.0
            
            # 1. Balance-based adjustment
            balance_mult = min(current_balance / 10.0, 3.0)
            multiplier *= balance_mult
            
            # 2. Performance-based adjustment
            if not self.recovery_mode:
                win_rate = self.calculate_win_rate(symbol)
                multiplier *= (1.0 + (win_rate - 0.5))
            
            # 3. Volatility adjustment
            volatility = self.get_market_volatility(symbol)
            if volatility > 1.5:
                multiplier *= 0.7
            elif volatility < 0.5:
                multiplier *= 1.2
            
            # 4. Compound level adjustment
            multiplier *= (1.0 + (self.compound_level - 1) * 
                         self.compound['base_increment'])
            
            # Calculate final lot size
            adaptive_lot = base_lot * multiplier
            
            # Apply safety limits
            min_lot = 0.001
            max_lot = (current_balance * 
                      self.pairs[symbol]['risk_settings']['max_position_size'])
            
            return round(max(min_lot, min(adaptive_lot, max_lot)), 3)
            
        except Exception as e:
            self.logger.error(f"Adaptive lot calculation error: {str(e)}")
            return base_lot

    def execute_trade(self, symbol: str, direction: str) -> bool:
        """Execute trade with enhanced safety checks"""
        try:
            # Pre-trade validations
            if not self.validate_trade_conditions(symbol):
                return False
            
            # Calculate position size
            base_lot = self.calculate_optimal_lot(symbol)
            final_lot = self.calculate_adaptive_lot(symbol, base_lot)
            
            # Get market data and calculate levels
            market_data = self.get_market_data(symbol)
            entry_price = float(market_data['last_price'])
            
            atr = self.calculate_atr(symbol)
            sl_distance = atr * self.pairs[symbol]['risk_settings']['stop_loss_atr']
            tp_distance = atr * self.pairs[symbol]['risk_settings']['take_profit_atr']
            
            sl_price = entry_price - sl_distance if direction == "long" else entry_price + sl_distance
            tp_price = entry_price + tp_distance if direction == "long" else entry_price - tp_distance
            
            # Execute order
            order = self.client.place_order(
                category="linear",
                symbol=symbol,
                side="Buy" if direction == "long" else "Sell",
                order_type="Market",
                qty=final_lot,
                stop_loss=str(round(sl_price, 4)),
                take_profit=str(round(tp_price, 4)),
                reduce_only=False,
                close_on_trigger=False
            )
            
            # Process successful trade
            if order['ret_code'] == 0:
                self.process_successful_trade(
                    symbol, direction, final_lot, 
                    entry_price, sl_price, tp_price
                )
                return True
            else:
                self.logger.error(f"Order failed: {order['ret_msg']}")
                return False
                
        except Exception as e:
            self.logger.error(f"Trade execution error: {str(e)}")
            return False

    [Rest of the implementation with additional methods...]
